add tags CRUD

+++ app/Http/Controllers/Admin/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
 use App\Models\Post;
+use App\Models\Tag;
 use Illuminate\Http\Request;
 
 class PostController extends Controller
@@ -14,7 +15,7 @@ class PostController extends Controller
      */
     public function index()
     {
-        $posts = Post::query()->with('category')->paginate();
+        $posts = Post::query()->with(['category', 'tags'])->paginate();
 
         $basket_cnt = Post::onlyTrashed()->count();
 
@@ -28,7 +29,9 @@ class PostController extends Controller
     {
         $categories = Category::query()->pluck('title', 'id')->all();
 
-        return view('admin.posts.create', compact('categories'));
+        $tags = Tag::query()->pluck('title', 'id')->all();
+
+        return view('admin.posts.create', compact(['categories', 'tags']));
     }
 
     /**
@@ -41,10 +44,13 @@ class PostController extends Controller
             'meta_desc' => ['max:255'],
             'content' => ['required'],
             'category_id' => ['required', 'exists:categories,id'],
+            'tags' => ['exists:tags,id'],
             'thumb' => ['max:255'],
         ]);
 
-        Post::query()->create($validated);
+        $post = Post::query()->create($validated);
+
+        $post->tags()->sync($request->tags);
 
         return redirect()->route('posts.index')->with('success', 'Post added successfully');
     }
@@ -66,7 +72,9 @@ class PostController extends Controller
 
         $categories = Category::query()->pluck('title', 'id')->all();
 
-        return view('admin.posts.edit', compact(['post', 'categories']));
+        $tags = Tag::query()->pluck('title', 'id')->all();
+
+        return view('admin.posts.edit', compact(['post', 'categories', 'tags']));
     }
 
     /**
@@ -81,10 +89,13 @@ class PostController extends Controller
             'meta_desc' => ['max:255'],
             'content' => ['required'],
             'category_id' => ['required', 'exists:categories,id'],
+            'tags' => ['exists:tags,id'],
             'thumb' => ['max:255'],
         ]);
 
         $post->update($validated);
+
+        $post->tags()->sync($request->tags);
 
         return redirect()->route('posts.index')->with('success', 'Post updated successfully');
     }
@@ -120,6 +131,10 @@ class PostController extends Controller
     public function basketDestroy(string $id)
     {
         $post = Post::withTrashed()->findOrFail($id);
+
+        if ($post->tags()->count()) {
+            return redirect()->route('posts.basket')->with('error', 'There are tags for this post');
+        }
 
         $post->forceDelete();
